Created Custom File Handler
Added Custom File Handle that will now create the directory if it does
not exist.

It will still fail if you go super deep like this:

../log/log/log.out , because the main directory (log) does not exist and
it will try to create ../log/log . Maybe that's something we can improve
on. However, I think it's not really a priority. It will create the
current folder by itself now though